refactor(app): rename the catalog page to collection
The page renders the org's persisted components (the collection), not the
shipped definitions (the catalog) — align it with the catalog/collection
taxonomy adopted by the agent. Renames the page, nav item, CollectionTable
component and useCollectionRows composable, plus the graph internals that
called the org graph "catalog" (useCollectionGraph). A stub page keeps
/catalog bookmarks working via a client-side redirect. The definitions
side (useCatalogStore, types/catalog, CatalogCard) is catalog-correct and
untouched.

By Digitl
